Make pure python versions of Javier's original notebooks
The aim is to allow easier tracking of differences between versions of
notebooks. I am using jupytext to pair each notebook with a pure
python version.  These are the commands that I have run:

$ jupytext --set-formats ipynb,py:light *.ipynb
$ jupytext --sync *.ipynb

I am hoping that I will now be able to rebase my modified notebook
branch on top of this branch, which should give nice simple diffs of
the python files associated with each notebook.